backend/requirements.txt
langchain-core
langchain-community
langchain-openai
langchain_google_genai
python-dotenv
datasets
transformers
torch
fastapi
uvicorn